backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
psycopg2-binary
python-dotenv
neo4j
httpx
